requirements.txt
pywsitest
pytest
pytest-rerunfailures
allure-pytest
allure-python-commons
websockets
pytest-asyncio
jsonschema
websocket_client